Moved notification-bar related tests to own test file + moved relevant ping-related util functions to utils
